+++ app/mission/imap/+mag/+app/+imap/+control/Field.m
@@ -38,7 +38,7 @@ classdef Field < mag.app.Control & mag.app.mixin.StartEndDate
 
             arguments (Input)
                 this
-                results (1, 1) mag.Instrument
+                results (1, 1) mag.imap.Instrument
             end
 
             arguments (Output)